Fix overflow bug in 2fa for send transaction
Need 64 bits to represent values in satoshis

+++ src/ga_auth_handlers.cpp
@@ -728,9 +728,9 @@ namespace sdk {
         }
 
         try {
-            const uint32_t limit = m_twofactor_required ? session.get_spending_limits()["satoshi"].get<uint32_t>() : 0;
-            const uint32_t satoshi = m_tx_details["satoshi"];
-            const uint32_t fee = m_tx_details["fee"];
+            const uint64_t limit = m_twofactor_required ? session.get_spending_limits()["satoshi"].get<uint64_t>() : 0;
+            const uint64_t satoshi = m_tx_details["satoshi"];
+            const uint64_t fee = m_tx_details["fee"];
             const uint32_t change_index = m_tx_details["change_index"];
 
             m_limit_details = { { "asset", "BTC" }, { "amount", satoshi + fee }, { "fee", fee },
